# frontend/src/api-client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "YAESandBox API (Public)",
    "description": "供前端和外部使用的公开 API 文档。",
    "version": "v1"
  },
  "paths": {
    "/api/atomic/{blockId}": {
      "post": {
        "tags": [
          "Atomic"
        ],
        "summary": "对指定的 Block 执行一批原子化操作。\r\n根据 Block 的当前状态，操作可能被立即执行或暂存。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "要执行操作的目标 Block 的 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "包含原子操作列表的请求体。",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAtomicRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAtomicRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAtomicRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作已成功执行，若为Loading状态则还额外暂存了一份。"
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Block 当前处于冲突状态 (ResolvingConflict)，需要先解决冲突。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "请求中包含无效的原子操作定义。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "执行操作时发生内部服务器错误。"
          }
        }
      }
    },
    "/api/manage/blocks/{blockId}": {
      "delete": {
        "tags": [
          "BlockManagement"
        ],
        "summary": "手动删除一个指定的 Block。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "要删除的 Block ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recursive",
            "in": "query",
            "description": "是否递归删除子 Block。默认递归删除，非递归可能导致奇奇怪怪的问题？",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "force",
            "in": "query",
            "description": "是否强制删除，无视状态。",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作已成功执行。"
          },
          "400": {
            "description": "不允许删除根节点或请求无效（例如 Block 有子节点但未指定 recursive=true）。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Block的当前状态不允许删除。现在只允许删除Idle或Error，说实在的有点意义不明，也许以后设计成什么都可以删除。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "执行操作时发生内部服务器错误。"
          }
        }
      }
    },
...
    "/api/Blocks/{blockId}": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "获取指定 ID 的单个 Block 的详细信息（不包含 WorldState）。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "要查询的 Block 的唯一 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回 Block 详细信息。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Blocks"
        ],
        "summary": "部分更新指定 Block 的内容和/或元数据。\r\n此操作仅在 Block 处于 Idle 状态时被允许。",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "要更新的 Block 的 ID。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "包含要更新的字段（Content, MetadataUpdates）的请求体。\r\n            省略的字段或值为 null 的字段将不会被修改（MetadataUpdates 中值为 null 表示移除该键）。",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlockDetailsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlockDetailsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlockDetailsDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "更新成功。"
          },
          "400": {
            "description": "请求体无效或未提供任何更新。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "未找到具有指定 ID 的 Block。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Block 不处于 Idle 状态，无法修改。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "更新时发生内部服务器错误。"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AtomicOperationRequestDto": {
        "required": [
          "entityId",
          "entityType",
          "operationType"
        ],
        "type": "object",
        "properties": {
          "operationType": {
            "minLength": 1,
            "type": "string",
            "description": "操作类型。必须是 \"CreateEntity\", \"ModifyEntity\", 或 \"DeleteEntity\" (不区分大小写)。"
          },
          "entityType": {
            "$ref": "#/components/schemas/EntityType"
          },
          "entityId": {
            "minLength": 1,
            "type": "string",
            "description": "操作目标实体的唯一 ID。不能为空或仅包含空白字符。"
          },
          "initialAttributes": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "description": "(仅用于 CreateEntity 操作)\r\n要创建的实体的初始属性字典。键是属性名，值是属性值。\r\n如果不提供，则实体以默认属性创建。",
            "nullable": true
          },
          "attributeKey": {
            "type": "string",
            "description": "(仅用于 ModifyEntity 操作)\r\n要修改的属性的键（名称）。",
            "nullable": true
          },
          "modifyOperator": {
            "type": "string",
            "description": "(仅用于 ModifyEntity 操作)\r\n修改操作符。预期值为 \"=\", \"+=\", \"-=\" 等表示赋值、增加、减少等操作的字符串。\r\n具体支持的操作符取决于后端实现 (YAESandBox.Core.State.Entity.OperatorHelper)。",
            "nullable": true
          },
          "modifyValue": {
            "description": "(仅用于 ModifyEntity 操作)\r\n修改操作的值。类型应与目标属性和操作符兼容。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用于 API 请求的单个原子操作的表示。\r\n定义了对 WorldState 中实体的创建、修改或删除操作。"
      },
      "BatchAtomicRequestDto": {
        "required": [
          "operations"
        ],
        "type": "object",
        "properties": {
          "operations": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AtomicOperationRequestDto"
            },
            "description": "要执行的原子操作请求列表。该列表不能为空。"
          }
        },
        "additionalProperties": false,
        "description": "包含用于批量执行的原子操作请求列表。"
      },
...
      "BlockDetailDto": {
        "required": [
          "blockId"
        ],
        "type": "object",
        "properties": {
          "blockId": {
            "type": "string",
            "description": "Block 的唯一标识符。",
            "nullable": true
          },
          "parentBlockId": {
            "type": "string",
            "description": "父 Block 的 ID。如果为根节点，则为 null。",
            "nullable": true
          },
          "statusCode": {
            "$ref": "#/components/schemas/BlockStatusCode"
          },
          "blockContent": {
            "type": "string",
            "description": "Block 的主要文本内容 (例如 AI 生成的文本、配置等)。",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "与 Block 相关的元数据字典 (键值对均为字符串)。",
            "nullable": true
          },
          "childrenInfo": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "该 Block 的直接子 Block 的 ID 列表。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用于 API 响应，表示单个 Block 的详细信息（不包含 WorldState）。"
      },
      "BlockStatusCode": {
        "enum": [
          "Loading",
          "Idle",
          "ResolvingConflict",
          "Error",
          "NotFound",
          "Deleted"
        ],
        "type": "string",
        "description": "表示 Block 的不同状态。"
      },
...
      "EntityType": {
        "enum": [
          "Item",
          "Character",
          "Place"
        ],
        "type": "string"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UpdateBlockDetailsDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "(可选) 要设置的新的 Block 内容。\r\n如果为 null，则不修改 BlockContent。",
            "nullable": true
          },
          "metadataUpdates": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "description": "(可选) 要更新或移除的元数据键值对。\r\n- Key: 要操作的元数据键。\r\n- Value:\r\n  - 如果为非 null 字符串: 添加或更新该键的值。\r\n  - 如果为 null: 从元数据中移除该键。\r\n如果整个字典为 null，则不修改 Metadata。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "用于通过 PATCH 请求部分更新 Block 的内容和元数据。\r\n任何设置为 null 的属性表示不修改该部分。"
      },
...
    }
  }
}